requirements.txt
appdirs==1.4.4
asgiref==3.2.5
astroid==2.5.2
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.8
django-allauth==0.41.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-tables2==2.3.3
filelock==3.0.12
idna==2.9
importlib-metadata==3.10.0
importlib-resources==5.1.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.0
pbr==5.5.1
pep8==1.7.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.5.0
pylint==2.7.4
python-decouple==3.4
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
stevedore==3.3.0
stripe==2.46.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
wrapt==1.12.1
zipp==3.4.1
